ath11k: Fix rssi in station dump and skip tx bitrate update for mgmt frames
Fix signal and signal average in station dump by using real
noise floor and appropriate bandwidth offset. In this way
signal strength in station matches the rssi_in_dbm in htt stats
Tx bitrate in station dump was dropping to legacy rate
randomly, this is because the tx rate was updated for
management packet as well. Skip tx bitrate update for
management packets based on the tid.

Change-Id: I68e82ca45665ae12e85e5307e6f8181f7a9a759c
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>