Mac80211: Initialize EWMA fail avg to 1
In certain scenario with more number of pass frames,
ewma_fail_avg would reach 0(lowest value) and in ewma method
value with 0 is a tricky and that could give some higher value
even with single failure value and that triggers MPL broken logic.
Note that it is already initialized to 1 during the time of init
or new mesh path assignment.

Change-Id: Id2034e4d40e1f4c002faf4f7678930778a3c2e3d
Signed-off-by: Tamizh Chelvam Raja <quic_tamizhr@quicinc.com>
Signed-off-by: Karthik M <quic_karm@quicinc.com>