Revert "ath12k: disable HW learning only for native wifi mode"
This reverts commit be3364c629d589529374852ee0518d77353fcd7d.

In mesh mode, driver receives additional peer map event with the same
peer_id. This is because of Firmware is doing the 4 address WDS
learning for mesh VAP and giving peer map event with same peer_id.

The 11s Mesh frames are not a WDS address frames, they are 6
address frames and the 4 address learning in firmware is not expected
for mesh frames. Because of these reasons, firmware recommends driver
to disable this HW WDS learning in all frame modes.

Change-Id: Ie571a705cb2df2e9bd1881ba73685f5f49199541
Signed-off-by: P Praneesh <quic_ppranees@quicinc.com>