Changes to fix compiltaion error for 32-bit architecture
Depending on the type of architecture the ARCH and the
K_DIR will be obtained in the buildtime and passed to the
make command accordingly.
Specifying LC_ALL = "C" will force the build environment to
follow the quilt series of applying patches.

Tested both 32 and 64 bit builds and tested AP bringup and
sanity.

Change-Id: I3138ae488692186420b2659ffb291cdbdb309171
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>