ath11k: Fix qcn9074 bring-up
During package upgrade, to debug qcn6122 bring-up issue we have added
the condition, which we have not removed after it worked.

For qcn6122 the internal_pci is 'true', so it's not hitting the return
statement, but for qcn9074 as we are not defining the value of internal_pci
in the hw_params, so by default it's taking 'false' and hitting the return
statement, which is causing issue in interrupts.

Fixes: 86a6ba117e ("mac80211: Package Upgrade 08/18/2022")

Change-Id: I424507d2bfa0668a07ab1d644ab7f875ba623f78
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>