Support for Agile DFS
Agile DFS feature can be used for background cac
in case of both FCC and ETSI domain. In case of
ETSI, it is known as precac whereas for FCC
it is known as rolling cac.

The changes are added to handle agile dfs and
it add support for simulating radar for adfs

Change-Id: Ia5fafc3b3bf57ae049f9cc6f7aa5c465714fc00f
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>