mac80211: WAR to fix warning during AWGN simulation in repeater
6 GHz has different power level for different modes and
hence channel context will be different for AP and STA
which is expected.

During issue, since there are 2 channel context for AP and STA,
warning is thrown since AP thinks it as channel context for
different channels.

Fixing the issue through WAR: removed the warning and will add
the validation once design is changed

Change-Id: I5eba7bfad050a334b38a03fce717f865b821bc77
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>