DFS interference bitmap support during DFS interference detection.
Currently when DFS interference is detected the whole configured
working frequency is moved to unavailable state. However the
target provides a way to identify the frequency where the
DFS interference is observed.

These changes calculate the DFS interference frequency based on the
offset received from target and generates the bitmap which denotes the
sub channels where DFS interference is detected. In the bitmap, the
LSB denotes the lowest 20MHz based on the center frequency configured.
Based on this bitmap the affected channels alone will be moved to
unavailable(NOL) state.

Also add support to get this bitmap advertized using a new attribute
"NL80211_ATTR_RADAR_BITMAP" with the event "NL80211_CMD_RADAR_DETECT".

Change-Id: Ib960576bdd7a091ee35b39bf47d9628a5659298a
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>