capability advertisement infra for multi-hw abstraction under one wiphy
There may be hw design (ath12k as of now) supporting MLO
across multiple discrete hw each acting as a link in the MLO
operation.
Since the prerequisite for MLO support in cfg80211/mac80211
is that all the links participating in MLO must be from
the same wiphy/ieee80211_hw.

Though most of the hw specific abstractions can be handled
with in the driver itself, there are some capabilities like
interface combination which can be specific to
each constituent physical hw.
This patch set tries to add an infrastructure to advertise
underlying hw specific capabilities like channel
and interface combinations.

Change-Id: I7ea49dce141a2bd68308837f0d6ab579e5dba149
Signed-off-by: Vasanthakumar Thiagarajan <quic_vthiagar@quicinc.com>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>