ath12k: setup ML reo queue LUT
Setup the ML reo queue look up table for handling
data from ML peers. The ML peer id created by the host
is used for managing this table.
Note that only the ML reoq is setup for ML peer and the
non ML LUT for the link peers is not required.

Change-Id: I1ecdf334e2a152a415027ac1e996912a19708e95
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>