mac80211: Add support to enable/disable FILS/UBPR frames
Add Support to enabled/disable the Fast initial Link Setup(FILS)
ands Unsolicited Broadcast Probe response frames in run time.

Currently FILS or UBPR config with interval 0 is considered
as invalid config, Allow user to set the fils max_interval and
ubpr interval to be set to 0 in which is case it is disabled.
This will allow run time disbable of FILS/UBPR rather doing
a BSS restart.

Change-Id: If82d0fdde42490907fcc8ddc98d07ed6247d4092
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>
Signed-off-by: Rameshkumar Sundaram <quic_ramess@quicinc.com>