ath12k: Fix vdev delete function
Currently, when bring up the monitor interface in 2g and make a switch
to 5g, it works, but subsequent try to switch it back to 2g, system
get stuck.

Internally infinite loop is getting introduced because while switching
to 5g, arvif->list is not getting removed from ar->arvifs, and trying
again to switch back to 2g, it is again adding arvif->list which is
same thing, and hence it is introducing cycle in the list, which leads to
infinite loop.

Fixed, by removing the arvif->list during deletion of vdev.

Change-Id: Ie563829f21753aaced448d465be40f4402b6a0e1
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>