cfg80211: Send wiphy index in AWGN NL Event
Currently, wiphy index is not passed in AWGN NL80211
msg. Due to this after wifi restart, when AWGN is detected,
hostapd fethces incorrect index which results in incorrect bss
and AWGN detection is run on 2G, 5G interface. This causes the
hostapd to bring down the interface since no channel is available.

Change-Id: Idc85731ece3f3c39bdaf88cf48427d2ea4d1e0b7
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>