ath12k: Fix unbalance in dma_map and dma_unmap
Currently, whenever a faulty address is getting generated during
dma_map_single, the faulty address is getting stored in the ti.paddr and
dma_unmap is happening for the faulty address, which is incorrect.

Due to this a crash is getting observed, whenever the above scenario
happens.

Fixed by adding a new label to avoid unmaping of a faulty address.

Fixes: 86a6ba117e ("mac80211: Package Upgrade 08/18/2022")

Change-Id: I2b5602e61753c85955bf8a981462cf425a75a273
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>