ath12k: Monitor mode bring up in SLO
Currently all the radios are sharing the same chandef, so during monitor
start wrong values are getting passed to FW, due to which we are seeing
FW assert.

Fixed by adding a check, when ar is not compatible with the chandef,
it will stop the process.

In addition, cfg80211 was not allowing to change the channels before
bringing up the interface, it needs at least one interface up to
change the channels.

Hence, removing the check for number of running interfaces.

Change-Id: I57988238880ade999f3a08dd41cbc2c91ac6e583
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>