ath12k: add new helper function ath12k_mac_is_ml_arvif()
With Multi-Link Operation (MLO), often there is a need to check whether
the link specific arvif is part of a MLD or not at many places.

Add ath12k_mac_is_ml_arvif() as a helper function which can be called
whenever there is a requirement to check the above mentioned condition.

Change-Id: I1366a2bb3998d4cb8035334b3486fa540d3f4faf
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>