Channel Switch support with MLO including Critical Update
Currently, Channel Switch is handled only for non-MLO case.

Add support for Channel Switch with MLO.

Also, Channel Switch classifies as Critical Update. Critical Update (CU) is
the explicit signaling for STA to notify any change in BSS parameters, so
that other STAs which are part of the same BSS can take the corresponding
action.

Add initial support for CU as well.

Change-Id: I23b8d8447b1a082be70c5194831c5c22f8c7ba68
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>