ath12k: improve iterations over channel contexts for Singly Wiphy
At various places, there is a need to iterate over channel contexts available
under the hardware. With Single Wiphy architecture, there will be multiple
channel contexts available from various underlying hardwares and hence, there
is a need to have the ar info in the iterator in order to decide whether the
current iterating channel context belongs to the ar under consideration or
not.

Add modification in the iterator functions to support multi-link operation.

Change-Id: Ie6d655661c6880b932110b5179035abcc6ed3472
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>