ath12k: Fix interrupt issue in Tx completion for 4th Tx ring
Interrupt mask gets set wrongly for the 4th Tx data ring which
uses WBM2SW4 RELEASE ring for completions. But since the ring
mask for the 4th ring points to WBM2SW3 currently,
interrupt mask is not set properly and as a result packets
sent on this ring are lost and results in OoM soon or later.

Fixes: I3a50428 "mac80211: Package Upgrade 08/18/2022"

Change-Id: I9fec567603f7ed84527b1bfbc24208f6a949a4c1
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>