ath12k: Fix Multiple issues in Puncturing bitmap configuration
The Preamble pucturing bitmap is a 16bit value where each bit
corresponds to a 20Mhz Subchannel(LSB is lowest 20Mhz).
Allow patterns that has only bits configured for current
operating BW.

The format that FW expects the puncture bitmap is inverse
of what is stored in Mac80211 chandef,(in chandef,
a bit set means the correpsondig 20Mhz sub-channel is Punctured,
but same is considered as enabled in FW)
During Multivdev restart puncture bitmap has to be inverted.
Also there is no dependency that puncture bitmap
shold be set only if puncturing support for HE is there,
hence remove the addtional check.

Change-Id: I0c9cfa4b4eced40894eb7c5ec74c1effd3a9c246
Signed-off-by: Rameshkumar Sundaram <quic_ramess@quicinc.com>