mac80211.sh: Handle race condition in device configuration
In case of MLO, the design is to start the single hostapd
instance when all the radio configs are available.

In this scenario, the driver setup was done simultaneously
for radio2G/radio5G/radio6G because of which the shared
file that was used to determine whether to start the
hostapd instance had a reader-writer problem.

To address this, leveraged the pre_wifi_down() from
openwrt to create a shared file which would mention
the number of radios up and number of mld vaps. Also,
instead of modifying the shared file, used a
tmp file instance to determine when to start the single
hostapd instance

Also, addressed following
   a.Interface naming is as follows
	1. for mlds - wlan0, wlan1, wlan2 ... wlanX
	2. for mld and legacy band B - wlan0, wlanB-1, wlanB-2 .. etc
		where B is 0,1,2 based on radioX_bandB config
	3. for sld - wlan0,wlan1,wlan2, wlan0-1, wlan1-1 .. etc - same as
	   legacy
   b.Do not update the config file if no mlds are present
   c.handle wifi down with pid for single wiphy case

Change-Id: I7afa9601c7b2c2edd9baf3ec507829cd5c66e5ce
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>