ath12k: fix Factory Test Mode crash issue
Currently we are getting ar by vif but for FTM vif is null
so we are getting null value for ar also. passing null ar
pointer in ath12k_tm_cmd.

we are converting hardware cookies for FTM mode
but firmware is not handling such type of cookie conversion,
so it's trigger the firmware assert.

Fix these issue by assigning ar value to ath12k_hw radio and
by disabling hardware cookie conversion in FTM mode.

Change-Id: I93703125c2d7c2508f477af281cb958de7bbd436
Signed-off-by: Aaradhana Sahu <quic_aarasahu@quicinc.com>