mac80211.sh: rdp454 single wiphy config support
In existing design, the config file and capabilities
are generated based on the band info present in the
phy information.

Since, in case of rdp454, 5G high and 5G low band
are advertised together. we might not able to generate
the device config for one of these bands.

To address this, add changes as follows
	1. while generating the wireless config file,
compare the channels of band and hw idx information
present in phy info. If the entire band channels does not
match with hw idx channels, then check if partial channel
list is matched. If yes, then generate a new device config
	2. while generating the hostapd config file, during
ht caps parse, get the band id that matches the hw idx channel

Change-Id: I415538409b86a071b5a7c7b4f4373c47d603d525
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>