mac80211.sh : fix multi mld issues
The code changes are done to address the following
	1. when mld config is enabled and reboot,
the mlds does not come up. This is because, the
shared file is saved in temporary directory which
gets removed on reboot. Hence, keep the shared
file in persistent directory
	2. when 2G +5G and 5G +6G mld is brought up,
wlan0 interface ifconfig is tried to bring up before
hostapd starts. Hence, it fails to start. Bring ifconfig
only if hostapd is started
	3. when false mld option is given in interface
config without wifi-mld, interface name is not wlanX-Y,
instead its wlanX. To keep in line with design handle the
same
	4. during wifi down, hostapd config files are
removed. No need to remove it because it can be used
for debugging

Change-Id: I5ab3158add71c6b80ce5352e4ee8279e763394cd
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>