ath12k: move config caching to ahvif
Currently the caching of arvif is done whenever the
arvif is not created, which happens during the channel
assignment time. But, the config can be cached since
arvif was preallocated and drv private of vif. But since
the drv priv is now ahvif and arvif is allocated dynamically
during channel assignement, the caching of parameters need
to be done at ahvif level for each link and the config
needs to be flushed to the vdev when arvif is binded to
the corresponding link.

Change-Id: Ie78bf03abf58ecc0f5fe7f7a3c733afc1695b175
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>