ath12k: fix debugfs dangling pointer
In rare scenario, remove link is invoked first
because of ieee80211_stop command possibly from
netdev_ops and then driver interface remove is invoked.
Because of this sequence, twt_debugfs pointer becomes a
dangling pointer because the parent directory is removed first.

This is reproduced only once and with effort.

Add changes to handle dangling pointer, remove
directory only if its parent directory is present.

Change-Id: I6ce6f13d4dbf94fa5fe471c40d89d18b2c3884e6
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>