ath12k: add support to get rssi value of non-associated client
Monitor the surrounding network to get RSSI of non-associated
client(i.e Neighbor-Peer associated to neighboring AP's).
Obtaining the RSSI it is useful for roaming between AP's
in a network and on deciding which AP the client should
associate with.
This is achieved through debugfs by providing the mac address
of a nrp which creates a neighbor-peer object to store's the
data associated with the nrp such as its mac, RSSI and timestamp.
Once the nrp is added a file with the mac-address of nrp as its
filename is created.
Each time the file is read a peer object is created for the
NRP and filter is set to get its ppdu_info for home channel
thereby obtaining the RSSI and a corresponding unix
timestamp is created, once the data collection is completed the
peer object is destroyed to free-up peer count.
Upto 6 NRPs can be created to keep the load and peer-count low.
Once the rssi and timestamp are obtained the peer object
created for NRP is deleted to free the peer id.
In case the NRP packets are unavailable during scan the
previous timestamped value is returned.

Command to add a Neighbor-Peer:
echo "add,<mac-addr-of-nrp>" >
/sys/kernel/debug/ieee80211/phyX/ath12k_hwY/neighbor_peer

Command to delete a Neighbor-Peer:
echo "del,<mac-addr-of-nrp>" >
/sys/kernel/debug/ieee80211/phyX/ath12k_hwY/neighbor_peer

Command to display RSSI of Neighbor-Peer:
cat /sys/kernel/debug/ieee80211/phyX/ath12k_hwY/nrp_rssi/<mac-addr-of-nrp>

Change-Id: Ic4c31d8bb9ca32ce5dce2db6d551b5b05b246053
Signed-off-by: Suresh Krishna <quic_sureshkr@quicinc.com>