ath12k: Fix search flag and update peer map event for sta
During STA throughput measurement, majority of packets getting drop due
to tx ring full error and it reduces the performance. This is because
of filling the incorrect ast_idx/ast_hash value to the tx_desc and vdev
search type/flag.

So that, add changes to correct the offset of ast_idx/ast_hash value
from peer map event and update vdev search type as ADDRX and enabled the
AST lookup in bank config to know the hw to use ast_idx/ast_hash from the
tx desc.

Change-Id: I7bb09049c05cc283385c69d15fc13ecb624c1fe7
Signed-off-by: Thiraviyam Mariyappan <quic_tmariyap@quicinc.com>
Signed-off-by: Karthikeyan Kathirvel <quic_kathirve@quicinc.com>