mac80211.sh: non-mld to mld copies ml information
when uci config file has mlo_capable=0 under device
config, still the mld configs are copied if the wifi-iface
has mld option.

When 2link mlo moved to 3 link mlo, still all the 3 links
are brought up because the config files are present

in case of SLO, wifi down does not remove the interface
from global instance. This is because the ifname check in
remove does not check the requested interfaces.

Add changes to handle the same

Change-Id: Iec28dae8b0dbaf28a9cbc0feb91127bf9ea51b1b
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>