ath11k : fix sta association in mbssid case
ath11k station is not able to associate to non-transmitting vif of mbss.

This is because wpa_supplicant does not receive the information about
non-transmitting vif as part of NL scan results.

ath11k driver does not advertises its mbssid capability to mac80211.
So wiphy->support_mbssid is not set.

Since support_mbssid is not set, the information about non-transmitting vif
is not parsed from the rx mgmt frame and it is not added as part of the scan
results.

Fix this by advertising the MBSSID capabitlity of ath11k driver to mac80211.

Fixes: I498845becb8 ("mac80211: re-add MBSSID/EMA support")

Change-Id: Ifcc5e68847e89049e4a8cb0b3c10430e48999c9f
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>