ath12k: improve hw scan handling in driver
When a scan request is received, driver selects a link id
for which the arvif can be mapped. Same link is also
used for getting the link conf address. Currently we return
0 as link id for a non ML vif, which is correct since that is
the default link id.
Also when any of the link vif is active and the scan request
is for a channel in the active link we return its link id.
But, when we dont hit both of the above cases(i.e not a ML
vif or no active link vif for the channel is present) we
currently return 0 as the link id.
This might not work out always, ex, when only one
link(ex. linkid = 1) is added to vif, then we wont find any
link conf for link id 0 in the vif resulting in scan failure.

During AP bringup, such scan failure can cause bringup issues.
Hence avoid sending link id 0 as default. Rather use a default
link for scan and default link address for the same. This scan
vdev will either be deleted if another scan is requested on same
vif or when AP is broughtup on same link or during interface cleanup

Also, scan vdev is created during hw scan start but deleted only when
channel is assigned. This also leads to FW assert during non-ML + ML
vif scenario.

Hence, address these issues via -
 * creating and deleting scan vdev during scan start and scan complete/abort
   respectively.
 * getting proper link id for creating scan vdev. If link_id can not be found
   then use last link id always for scanning purpose instead of returning 0.

Change-Id: I68c45dee09c35393ff34c2b9bfce9ba6c49e84d7
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>
Co-Developed-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>