ath12k: fix 320Mhz bytes/packets counters in ext_rx_stats
In case of 320MHz, for both 320MHz_1 and 320MHz_2 channelization,
where FW updated 4 and 5 respectively, bandwidth index 4 is used
internally in ath12k driver. Rx Stats are aggregated in index 4
but during display, index 5 is used.

Changing the stats dump to use right index to display the stats for
320MHz.

Fixes:  I2a6f43e5a45 (ath12k: Modify ppdu info bandwidth in u_sig_cmn
TLV)
        Iba35b4e1ab8 (ath12k: fix 320Mhz bytes/packets counters in peer
stats)

Change-Id: Id86f67597174394b30c733964c3ca3ce87450636
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>