mac80211.sh: single wiphy optimize vap clean up
currently, vap clean up is tightly coupled with hostapd
interface names. During clean up in the script
for single wiphy case with ML vaps we would take interface
name list from the hostapd config file because when
ML and non-ML vaps was configured, there was a inconsistency
in removing the interface. This method of clean up might not be scalable.

Hence, add changes to add the interface names to wifi-device
which can be used during cleanup.

Change-Id: I391c937ca5e3f639c1ae8b9971b66bbe0560e494
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>