ath12k: Fix NULL link arvif access during tx
The link arvif gets assiged and unassigned during channel
assign and unassign. There could be race when the arvif could
be cleared while Tx gets prepared for that link via a mgmt Rx which
requires Tx after Rx processing (Ex. invalidation of action frames)
In such cases, the arvif could be NULL, but due to missing NULL validation,
below crash is seen.

[ 3721.415953] Unable to handle kernel read from unreadable memory at virtual address 0000000000000018
[ 3721.415973] Mem abort info:
[ 3721.423791]   ESR = 0x96000005
[ 3721.426569]   EC = 0x25: DABT (current EL), IL = 32 bits
[ 3721.429688]   SET = 0, FnV = 0
[ 3721.435158]   EA = 0, S1PTW = 0
[ 3721.438018] Data abort info:
[ 3721.441060]   ISV = 0, ISS = 0x00000005
[ 3721.444182]   CM = 0, WnR = 0
<snip>
[ 3721.765378] pc : ath12k_mac_get_any_chandef_iter+0xaf4/0x36d8 [ath12k]
[ 3721.770153] lr : ieee80211_nullfunc_get+0xbec/0xd30 [mac80211]

Hence, make the NULL validation of arvif common for ML and non ML vif
and drop frames when it tx is requested on invalid link.

Fixes:  Ib635d5f6b416f ("ath12k: avoid using deflink by default during tx op")

Also, moving to RCU for arvif access is recommended as followup
to make access clean and safe.

Change-Id: I11d0aecec8f5f76bc70545e9ea89126982c31021
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>