ath12k: Add support for mlo in cnss diag debug
Currently, when mlo is enabled when we run cnss_diag -p -c
radio index is always 0 (R0: FWMSG:...) due to this it is
not possible to identify radio and debugging is not possible.

Add ATH12K_LINK_IDX attribute to send ar->link_idx that is
fetched by cnss diag to display correct radio number.

Change-Id: Idb6db41208f8d402f4963cf917fdfd69b33a4438
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>