ath12k: reorder ath12k_exit() calls
Crash is observed in ath12k_debugfs_soc_destroy() during 'rmmod ath12k'
due to incorrect order of deinitializations.

The directory for debugfs_soc is created as a sub-directory of
debugfs_ath12k. However ath12k_exit() removes the latter first
recursively and then tries to remove the former resulting in a crash.

Reorder the deinitialization functions to follow the order opposite
to the initializations.

Change-Id: I1270f6bc5a7425e1032859bfda54872062588feb
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>