ath12k: fix firmware assert during insmod in memory segment mode
Firmware allocate each segment of size 2MB for memory segment mode
which is calculated via num peers value send by host to firmware
during wmi init.

Currently, num peers value is 512+17 but in case of split phy this
value become 2*(512 + 17) which is 1058 due to this it's trigger
firmware assert.

So, Fix this issue by reducing num_peers value from 512 to 128.

Change-Id: I87d724930373be96ebc69428552af94a379aaf66
Signed-off-by: Aaradhana Sahu <quic_aarasahu@quicinc.com>