ath12k: remove WARN_ON from channel update callback
Currently, when channel update comes to driver and if the interface is not
started, WARN_ON is thrown. However, with MLO core driver changes, there
can be a cases where vdev is waiting for its partner vdevs to get created
first before starting itself.

Hence, in this situation it is possible to hit the above situation.

Remove WARN_ON since the situation is expected.

Change-Id: I0e007c88b128f78aa125e12dd8684be961924f41
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>