cfg80211: handle DFS cancel notifying in link clearing case
When an active CAC is going on and interface is brought down, currently
while clearing the link it will try to send netlink notification that
CAC got aborted. However, before calling the notifier, the wdev link bitmap
will be cleared and in the notifier, we have a check that the link map
should be set. This leads to a condition that kernel dont send out the
notification and reset the CAC state back. And during final interface
stop, due to wrong state, WARN_ON is thrown.

In the notifier we already have a check that if the corresponding link
doesnt have CAC active, dont proceed with any notification. Hence it is
safe to remove the wdev bitmap set logic.

Remove logic to check link bitmap in wdev during DFS notification.

Change-Id: I3268009a98fe038ba4920ef5c75ddd3c680efc5a
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>