Update link id after op supports
arsta ptr is NULL within ath12k_mac_op_sta_rc_update()
though sta and arvif is valid.

If station is associated on link1 and link id 0 is
hardcoded in ath12k_mac_op_sta_rc_update and we are
checking arsta at link id 0 which is NULL due to this
crash is observed.

Handling rc update for STA based on link id needs to be supported.

Change-Id: I1ae2df6e51fb6b515a865e0e54a293fbbb2c1410
Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>