mac80211: Fix association failure in case of mlo 5G 240 MHz with prop sta
The channel info from default bss conf in sdata may not be valid in case of
MLO enabled and accessing the band info from bss conf channel causes the below
crash in mac80211.

[  350.522065] Unable to handle kernel access to user memory outside uaccess routines at virtual address 0000000000000000
[  350.977531] Call trace:
[  350.982837]  ieee80211_eht_cap_ie_to_sta_eht_cap+0x104/0x1e4 [mac80211]
[  350.985011]  ieee80211_nan_func_match+0x4f0/0x3a24 [mac80211]
[  350.991607]  ieee80211_channel_switch+0x109c/0x1f9c [mac80211]
[  350.997510]  ieee80211_channel_switch+0x1570/0x1f9c [mac80211]
[  351.003235]  __cfg80211_send_event_skb+0x23ac/0x43c8 [cfg80211]

Since sband is already passed in the function use it for 5G band validation
to avoid the issue.

Change-Id: Ic034aa5ff48fa3c1d93c87c67629f16fac8a6463
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>