ath12k: fix tcl_metadata vdev_id info
During AP-STA repeater scenario, initially,
AP interface is brought up with vdev_id 0 and
in parallel, STA interface is brought up with
vdev_id 1 and scan is initiated for STA interface.
While the scan is on going, AP interface is disabled
as the STA is not yet associated. Hence, vdev_id 0 is
deleted.

Once STA scan is completed, vdev (vdev_id 1) is deleted
and during association again STA interface is created with
vdev_id 0. During this, the vdev_id information is appended
in tcl_metadata which results in incorrect configuration of
vdev_id because the ahvif's deflink still had the old vdev_id
in the tcl_metadata information.

Hence, add changes to write tcl_metadata config instead of
appending it.

Change-Id: I6e93f242c70feb2e5f96a8d5ba5dad61f5439fd5
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>