ath12k: fix filling parameters of WMI TPC command
Currently, if TPE IE is not present as well as channel is not
a PSD channel, the tx power is taken as minimum of eirp power
and the regulatory TPE IE power. But, if TPE IE is not present,
regulatory TPE IE will be 0 and hence min will always return 0.
Add fix to use directly the eirp power when TPE IE is not present.
Fixes: I5eb8f6e5b25(ath12k: Regulatory support for 6 GHZ with new
client)

Change-Id: I05aeba5c322279e886f1581193d9c3c9f3ea0a00
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>