cfg80211:  fix 6GHz AP-STA concurrency issue
Currently, for 6 GHz band, 9 different channel pools are maintained
based on the various 6G regulatory power mode a device can be
configured in. Due to this, chandef chan pointer will not be same
for AP and STA. During, AP-STA concurrency, chandef compatibility is
matched and for 6 GHz it fails which ultimately leads to nl80211_start_ap
returning EBUSY.

This changes were removed as part of refactoring the cfg80211 to handle
multiple hw abstraction. Hence, add an api to check the channel center_freq
information for 6 GHz bands

Also, these check are needed for multi-hw compatibility check as well.

Fixes: 830c2dd38c0f ("capability advertisement infra for multi-hw abstraction under one wiphy")

Change-Id: Iae5c344765fe9f748fa848b25b4a9fa80b617497
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>