mac80211: update radar_required in channel context after  channel switch
When the channel switch has been made, a vif will use the
channel context which was reserved. Channel context and channel
definition in sdata will also be updated accordingly.
In this case, radar required parameter passed from NL stored as
reserved_radar_required should also be updated to sdata.

When channel is switched from 80 MHz to 160 MHz where primary 80 MHz
is non DFS and secondary 80 MHz channels are DFS channels, radar
required in channel switch parameters will be true but it is not
updated to sdata when using reserved channel context.

This flag is used by ath12k driver to inform that the channel is radar
enabled to firmware during vdev start/restart.
Hence Update the radar required flag to fix the same.

Change-Id: I4b529e532e67def505e78c5e9bb6687cba8c7a2f
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>